make Profile Controller
