+++ src/test/java/com/threeNerds/basketballDiary/mvc/service/TeamMemberServiceTest.java
@@ -1,5 +1,6 @@
 package com.threeNerds.basketballDiary.mvc.service;
 
+import com.threeNerds.basketballDiary.mvc.controller.ProfileController;
 import com.threeNerds.basketballDiary.mvc.controller.UserController;
 import com.threeNerds.basketballDiary.mvc.repository.TeamMemberRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -27,7 +28,7 @@ class TeamMemberServiceTest {
 
     @BeforeEach
     void setUpEach(){
-        UserController.FindMyTeamProfileDTO findMyTeamProfileDTO = new UserController.FindMyTeamProfileDTO()
+        ProfileController.FindMyTeamProfileDTO findMyTeamProfileDTO = new ProfileController.FindMyTeamProfileDTO()
                 .userSeq(1L)
                 .teamSeq(3L);
